More fixes

+++ src/math/vector.hpp
@@ -113,7 +113,7 @@ public:
 
   Vector to_int_vec() const
   {
-    return Vector(floor(x), floor(y));
+    return Vector(floorf(x), floorf(y));
   }
 
   // ... add the other operators as needed, I'm too lazy now ...
